packages/server/pyproject.toml
[project]
name = "comfydock_server"
version = "0.3.2"
description = "ComfyDock Server"
readme = "README.md"
authors = [
    { name = "akatz-ai", email = "katzfeyalex+akatzgh@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "comfydock-core>=0.2.2",
    "fastapi[all]>=0.115.8",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
